Move utilities classes to core package
