+++ src/main/java/com/github/dockerjava/core/util/guava/Guava.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.dockerjava.jaxrs.util.guava;
+package com.github.dockerjava.core.util.guava;
 
 public class Guava {
