+++ src/main/java/com/github/dockerjava/core/util/guava/PercentEscaper.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.github.dockerjava.jaxrs.util.guava;
+package com.github.dockerjava.core.util.guava;
 
 import jersey.repackaged.com.google.common.base.Preconditions;
